Split dataset using inheritance
